Move validator/server_api/json.go to util/jsonapi/validation.go
