+++ util/jsonapi/validation_test.go
@@ -1,4 +1,7 @@
-package server_api
+// Copyright 2023, Offchain Labs, Inc.
+// For license information, see https://github.com/OffchainLabs/nitro/blob/master/LICENSE
+
+package jsonapi
 
 import (
 	"encoding/json"
